+++ tasks/adapt.py
@@ -64,7 +64,7 @@ def main():
 
     print(colored(f"Output file: {output_file}", "yellow"))
     if end_turn:
-        coa.split_scratchpad_output(output_file)
+        coa.split_scratchpad_output(output_file, task_settings["document_tag"])
         coa.run_latexdiff(args.input_file, output_file, args.task)
 
     coa.log_output_files(output_file, log_file)
@@ -88,7 +88,7 @@ def main():
         coa.log_output_files(output_file_reflect, log_file)
         coa.log_and_print_statistics(state, args.model, log_file)
         if end_turn_reflect:
-            coa.split_scratchpad_output(output_file_reflect)
+            coa.split_scratchpad_output(output_file_reflect, task_settings["document_tag"])
             coa.run_latexdiff(args.input_file, output_file_reflect, args.task)
             coa.run_latexdiff(output_file, output_file_reflect, args.task, args.model)